I made a new TouristPathSimulator named "TouristPathSimulatornew" with adding two linksampllers, one is the original Piter did and one is for the angle different and the link value.
In the end i made a If that give me the option with what probability to use the first or the second linksampller.

As well i made a new function in RandomCollecation named "selectRandom" that give me the probability of using a Random link pick or Minimum angle link pick.

i made  a new TouristPathSimulator named "TouristPathSimulatortest" with adding one linksampllers, angle different and the link value.